// src/helpers/typescript-helpers/enums.ts
export enum ReqBodyParts {
  BODY = "body",
  PARAMS = "params",
  QUERY = "query",
}

export enum Categories {
  PROPERTY = "property",
  TRANSPORT = "transport",
  WORK = "work",
  ELECTRONICS = "electronics",
  BUSINESS_AND_SERVICES = "business and services",
  RECREATION_AND_SPORT = "recreation and sport",
  FREE = "free",
  TRADE = "trade",
}